Completing per-class reference ignores
Tested on default TPP character BPs which hard reference SKM and ABP
assets. Excluding these brings the character down to a memory-reference
weight of ~4KiB. My test asset that references these TPP character by
way of using it's type in a soft-reference type is detected as having a
heavy reference weight as it should not be loading those assets.